requirements.txt
geopandas
descartes
osmnx
plotly_express
currencyconverter
python-dotenv